# swagger/swagger.yml
openapi: 3.0.0
info:
  title: TS Seed API
  description: Typescript + Seed API Project
  version: 1.0.0
servers:
  - url: https://ts-seed-api.com
    description: Main production server
  - url: https://ts-seed-api-staging.com
    description: Internal staging server

paths:
  /sample:
    parameters:
      - $ref: '#/components/parameters/DefaultLangHeader'
    get:
      tags:
        - Sample
      summary: GET Sample
      description: Get all samples in DB
      responses:
        200:
          description: A list of samples
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: 'components/sample.yml#/components/schemas/Sample'
        500:
          $ref: '#/components/responses/InternalError'
    post:
      tags:
        - Sample
      summary: POST Sample
      description: Create a sample
      requestBody:
        $ref: 'components/sample.yml#/components/requestBodies/PostSampleBody'
      responses:
        201:
          $ref: '#/components/responses/ResourceCreated'
        400:
          $ref: '#/components/responses/BadRequest'
        500:
          $ref: '#/components/responses/InternalError'
  /sample/{id}:
    parameters:
      - $ref: '#/components/parameters/DefaultUUIDPath'
      - $ref: '#/components/parameters/DefaultLangHeader'
    get:
      tags:
        - Sample
      summary: GET Sample by ID
      description: Get a specific sample based on its ID
      responses:
        200:
          description: A specific sample
          content:
            application/json:
              schema:
                $ref: 'components/sample.yml#/components/schemas/Sample'
        400:
          $ref: '#/components/responses/BadRequest'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalError'
    put:
      tags:
        - Sample
      summary: PUT Sample
      description: Update a sample
      requestBody:
        $ref: 'components/sample.yml#/components/requestBodies/PutSampleBody'
      responses:
        200:
          $ref: '#/components/responses/ResourceCreated'
        400:
          $ref: '#/components/responses/BadRequest'
        500:
          $ref: '#/components/responses/InternalError'
    delete:
      tags:
        - Sample
      summary: DELETE Sample by ID
      description: Delete a specific sample based on its ID
      responses:
        200:
          description: Returns an empty object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyBodyGeneral'
        400:
          $ref: '#/components/responses/BadRequest'
        500:
          $ref: '#/components/responses/InternalError'

tags:
  - name: Sample

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  responses:
    GenericSucces:
      description: A message saying everything went well
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BodyGeneral'
    ResourceCreated:
      description: A message saying everything went well with the id of the resource created
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/BodyGeneral'
              - type: object
                properties:
                  id:
                    type: string
                    format: uuid
    BadRequest:
      description: A message with the details of the bad parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BodyGeneral'
    Unauthorized:
      description: A message requesting an auth method
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BodyGeneral'
    NotFound:
      description: A message with the details of the item(s) not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BodyGeneral'
    Duplicated:
      description: A message describing the duplicated item in DB
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BodyGeneral'
    InternalError:
      description: A message with a generic message
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BodyGeneral'
  parameters:
    DefaultUUIDPath:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
      description: UUID of a resource
    DefaultLangHeader:
      in: header
      name: HTTP_ACCEPT_LANGUAGE
      required: true
      schema:
        type: string
        enum:
          - fr
          - en
      description: Language to use during the treatment of the request
  schemas:
    BodyGeneral:
      type: object
      properties:
        message:
          type: string
    EmptyBodyGeneral:
      type: object